+++ src/components/darkmode/style.css
@@ -1,0 +1,10 @@
+.dark-mode-area {
+  width: 100%;
+  background: var(--white);
+  padding: 10px;
+  box-sizing: border-box;
+  display: flex;
+  align-items: center;
+  justify-content: end;
+  border-bottom: 1px solid var(--border-color);
+}
